requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
flake8==7.0.0
frozenlist==1.4.1
idna==3.6
isort==5.13.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
python-dotenv==1.0.1
typing_extensions==4.10.0
yarl==1.9.4
